Use Riot to run benchmarks (#1915)
* Use Riot to run benchmarks

* run benchmarks in parallel

* do not capture coverage for benchmarks

* skip coverage for benchmarks
