+++ riotfile.py
@@ -39,6 +39,12 @@ venv = Venv(
                 "pygments": latest,
             },
         ),
+        Venv(
+            name="benchmarks",
+            pys=select_pys(),
+            pkgs={"pytest-benchmark": latest, "msgpack": latest},
+            command="pytest --no-cov {cmdargs} tests/benchmarks",
+        ),
         Venv(name="tracer", command="pytest tests/tracer/", venvs=[Venv(pys=select_pys(), pkgs={"msgpack": latest})]),
         Venv(
             name="pymongo",
