+++ tests/benchmarks/test_tracer.py
@@ -90,24 +90,3 @@ def test_tracer_large_trace(benchmark, tracer):
 
 def test_tracer_start_span(benchmark, tracer):
     benchmark(tracer.start_span, "benchmark")
-
-
-@pytest.mark.benchmark(group="span-id", min_time=0.005)
-def test_rand64bits_no_pid(benchmark):
-    from ddtrace.internal import _rand
-
-    benchmark(_rand.rand64bits, False)
-
-
-@pytest.mark.benchmark(group="span-id", min_time=0.005)
-def test_rand64bits_pid_check(benchmark):
-    from ddtrace.internal import _rand
-
-    benchmark(_rand.rand64bits)
-
-
-@pytest.mark.benchmark(group="span-id", min_time=0.005)
-def test_randbits_stdlib(benchmark):
-    from ddtrace.compat import getrandbits
-
-    benchmark(getrandbits, 64)
