+++ tox.ini
@@ -134,7 +134,6 @@ envlist =
     py27-opentracer_gevent-gevent{10}
     py{27,35,36}-opentracer_gevent-gevent{11,12}
     py{37,38}-opentracer_gevent-gevent{13,14}
-    benchmarks-py{27,35,36,37,38,39}
 
 isolated_build = true
 
@@ -169,7 +168,6 @@ deps =
     mock
 # used to test our custom msgpack encoder
     integration: msgpack
-    benchmarks: pytest-benchmark
     profile: pytest-benchmark
     profile-minreqs: protobuf==3.0.0
     profile-minreqs: tenacity==5.0.1
@@ -486,7 +484,6 @@ commands =
 # run subsets of the tests for particular library versions
     ddtracerun: pytest {posargs} --no-cov tests/commands/test_runner.py
     test_logging: pytest {posargs} tests/contrib/logging/
-    benchmarks: pytest {posargs} tests/benchmark.py
 
 [testenv:wait]
 skip_install=true
